Re-add epiphany demo
This has some caveats to be aware of
1. No idea if it would work on apple or windows OS as they are not linux
and would also need an X Server
2. Image is fixed to Ubuntu 22.04. Can't update the base image it won't
work.
3. Tested on a Ubuntu 24.04 VM and Raspian Bookworm 64bit on a raspberry
pi.

In general running GUIs from containers is complex requiring various
other components such as dbus to be running. In the case here it manages
to work with an environment variable hack, but moving to 24.04
introduced a new issue with dbus. KDE apps may also have issues. Even X
Can have issues like the shared memory extension. There are interesting
tools such as X11Docker which try to provide easier solutions but on the
whole trying to run gui applications from a container can be fragile.

Signed-off-by: Dave Kelsey <d_kelsey@uk.ibm.com>